requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-google-genai
faiss-cpu
PyPDF2
python-dotenv
langchain-text-splitters
